WIP: split metadata validation and tests (checkpoint)
